Split widget text fallback
